requirements.txt
python-telegram-bot==20.0
yfinance==0.2.18
pandas==2.1.4
beautifulsoup4==4.12.3
requests==2.31.0
pytz==2024.1